- Created a move and shoot effect which performs a shot effect at some point during a movement effect - Modified the BuffaloRider to use this MoveShootEffect as to have a seamless animation on the client side when ramming into targets - Added type safety for paths generated using the AStarPathUtil

+++ src/java/com/threerings/bang/game/data/effect/ShotEffect.java
@@ -105,7 +105,7 @@ public class ShotEffect extends Effect
     public String defendIcon;
 
     /** Direction location the target is pushed. */
-    public int pushx = -1, pushy = -1;
+    public short pushx = -1, pushy = -1;
     
     /**
      * Constructor used when creating a new shot effect.
